refactor(model): changed GooglePlayApp to App
- moved GoogePlayApp to the package "model"
- renamed GooglePlayApp to just App
- some minor nits
