+++ src/main/kotlin/com/appdetex/sampleparserjavaproject/parsing/JsonSerializer.kt
@@ -1,5 +1,6 @@
-package com.appdetex.sampleparserjavaproject
+package com.appdetex.sampleparserjavaproject.parsing
 
+import com.appdetex.sampleparserjavaproject.model.App
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.encodeToString
 
@@ -7,7 +8,7 @@ object JsonSerializer {
     // https://github.com/Kotlin/kotlinx.serialization/blob/master/docs/json.md
     private val prettyFormat = Json { prettyPrint = true }
 
-    fun asJson(app: GooglePlayApp): String {
+    fun asJson(app: App): String {
         return prettyFormat.encodeToString(app)
     }
 }
